+++ Changes.md
@@ -8,10 +8,19 @@ Improvements
 - Windows : Gaffer now uses the TBB memory allocator for significantly better performance.
 - ColorChooser : Changed the color field widget to a color wheel when hue is one of the varying components. [^1]
 
+API
+---
+
+- GafferCycles :
+  - Refactored bindings so they are no longer dependent on linking to Cycles.
+  - The `devices`, `nodes`, `shaders`, `lights`, and `passes` Python attributes now contain IECore.CompoundData instead of Python dictionaries.
+  - Added `majorVersion`, `minorVersion`, `patchVersion`, and `version` Python attributes containing the Cycles version.
+
 Breaking Changes
 ----------------
 
 - InteractiveRenderTest : Removed `interactiveRenderNodeClass` member and `useNodeClass` argument to `_createInteractiveRender`. All testing is now performed with the InteractiveRender node itself.
+- GafferCycles : The `devices`, `nodes`, `shaders`, `lights`, and `passes` Python attributes now contain IECore.CompoundData instead of Python dictionaries.
 
 1.5.0.0a3 (relative to 1.5.0.0a2)
 =========